requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Temporarily commented out - install separately if needed

# AI/ML
openai==1.9.0
anthropic==0.16.0
transformers==4.37.0
torch>=2.6.0
sentence-transformers==2.3.1
numpy>=1.26.0
scipy>=1.12.0

# YouTube and media processing
yt-dlp==2024.1.0
youtube-transcript-api==0.6.2

# Text processing
tiktoken==0.5.2
langdetect==1.0.9
pypdf2==3.0.1
beautifulsoup4==4.12.3

# Caching and async
redis==5.0.1
celery==5.3.6
aiohttp==3.9.3

# Utils
pydantic==2.5.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0